carprice-scraper/requirements.txt
playwright==1.45.0
requests==2.31.0
python-dotenv==1.0.1
pika==1.3.2
PyYAML==6.0.2
pandas==2.3.3
pandas-stubs==2.3.2.250926
